requirements.txt
altgraph==0.16.1
aniso8601==3.0.2
asn1crypto==0.24.0
autoenv==1.0.0
certifi==2018.4.16
cffi==1.11.5
click==6.7
cryptography==2.1.4
dis3==0.1.2
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.6
future==0.16.0
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
jeepney==0.3.1
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
macholib==1.10
MarkupSafe==1.0
mysql-connector==2.1.6
pefile==2017.11.5
pipenv==2018.7.1
psutil==5.4.2
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
PyInstaller==3.3.1
python-dotenv==0.9.0
pytz==2018.5
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
